renamed encoders namespace to serializers
