+++ src/Message/Serializers/AvroEncoder.php
